requirements.txt
asgiref==3.2.10
Django==3.0.8
django-ratelimit==3.0.0
fastapi==0.58.1
pydantic==1.5.1
pytz==2020.1
sqlparse==0.3.1
starlette==0.13.4
